refactor: move post info into shared
