+++ packages/maskbook/src/social-network/PostInfo.ts
@@ -1,10 +1,1 @@
-import { PostInfo } from '@dimensiondev/mask-plugin-infra'
-export { PostInfo } from '@dimensiondev/mask-plugin-infra'
-
-export const emptyPostInfo: PostInfo = new (class extends PostInfo {
-    commentBoxSelector = undefined
-    commentsSelector = undefined
-    rootNode = undefined!
-    rootNodeProxy = undefined!
-    postContentNode = undefined
-})()
+export { PostInfo, emptyPostInfo } from '@dimensiondev/mask-plugin-infra'
