+++ packages/maskbook/src/social-network/defaults/inject/PostInspector.tsx
@@ -4,7 +4,7 @@ import type { PostInfo } from '../../PostInfo'
 import { createReactRootShadowed } from '../../../utils/shadow-root/renderInShadowRoot'
 import { PostInspector, PostInspectorProps } from '../../../components/InjectedComponents/PostInspector'
 import { makeStyles } from '@material-ui/core'
-import { PostInfoContext } from '../../../components/DataSource/usePostInfo'
+import { PostInfoProvider } from '../../../components/DataSource/usePostInfo'
 import { noop } from 'lodash-es'
 
 export function injectPostInspectorDefault<T extends string>(
@@ -26,7 +26,7 @@ export function injectPostInspectorDefault<T extends string>(
     const zipPostF = zipPost || noop
     return function injectPostInspector(current: PostInfo, signal: AbortSignal) {
         const jsx = (
-            <PostInfoContext.Provider value={current}>
+            <PostInfoProvider post={current}>
                 <PostInspectorDefault
                     onDecrypted={(typed, raw) => {
                         current.decryptedPostContent.value = typed
@@ -35,7 +35,7 @@ export function injectPostInspectorDefault<T extends string>(
                     zipPost={() => zipPostF(current.rootNodeProxy)}
                     {...current}
                 />
-            </PostInfoContext.Provider>
+            </PostInfoProvider>
         )
         if (config.render) {
             const undo = config.render(jsx, current)
